* commit '7a5f23f287539d8e8d3bbceba9b169b6f8f58987':
  fixed some warnings from the merge
  changes from review feedback
  added marco around saveing out debug info
  fixed the conflict resolution bug
  made lookup function inlined, not sure it makes a difference as a callback function
  three way merge seems to be working, but the common transient part doesn't seem to be doing so well... I think there is a bug still

+++ raylib/rayutils.h
@@ -71,7 +71,7 @@ inline double random(double min, double max)
   return min + (max - min) * (double(rand()) / double(RAND_MAX));
 }
 
-inline std::vector<int> voxelSubsample(const std::vector<Eigen::Vector3d> &points, double voxel_width)
+inline std::vector<int64_t> voxelSubsample(const std::vector<Eigen::Vector3d> &points, double voxel_width)
 {
   struct Vector3iLess
   {
@@ -84,7 +84,7 @@ inline std::vector<int> voxelSubsample(const std::vector<Eigen::Vector3d> &point
       return a[2] < b[2];
     }
   };
-  std::vector<int> indices;
+  std::vector<int64_t> indices;
   std::set<Eigen::Vector3i, Vector3iLess> test_set;
   for (int64_t i = (int64_t)points.size()-1; i >= 0; i--)
   {
